changes in Motor_Test_Sketch
used LM339 comparator, for the MOSFET switching, and detection of 3 phase, and low voltage cutoff.

remaining to add ground plane, and output of the comparator for the switching of the mofet is yet  to tied.
added bypass capacitors for the input of the 3 phase detectoin, as well for motor feedback,
added input bypass caps, in 3 phase detection, and low voltage cutoff, as stated in the Texas Intruments Guideline for reducing EMI.